Sort column result order based on the columns option order
